y2-t3_keep-playing/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "keep-playing"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "Django==5.2.1",
    "djangorestframework==3.15.2",
    "django-environ==0.12.0",
    "gunicorn==23.0.0",
    "psycopg2-binary==2.9.10",
    "whitenoise==6.9.0",
    "Pillow==11.2.1",
    "python-docx==1.1.2",
    "lxml==5.4.0",
    "boto3>=1.35.0",
    "django-storages>=1.14.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-django>=4.9"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "*_test.py"
python_classes = "*Test"
python_functions = "test_*"
testpaths = ["tests"]
